[问题/需求描述]:取消邮件发送（有莫名其妙的问题） [解决方法]:取消邮件发送 [影响范围]:无 [需求编号]:无

+++ infra/recer/ctp_recer.h
@@ -454,7 +454,7 @@ class CtpTraderSpi : public CThostFtdcTraderSpi {
 
   bool front_disconnected = false;
   int re_connect = 0;
-  int session_id = 0;
+  uint64_t session_id = 0;
   int front_id = 0;
   std::string user_id;
   std::string req_transaction_cost_exchange;
